rename db folder to storage
