+++ config/config.go
@@ -5,9 +5,9 @@ import (
 	"path/filepath"
 	"strings"
 
-	storage "github.com/0xPolygonHermez/zkevm-synchronizer-l1/db"
 	"github.com/0xPolygonHermez/zkevm-synchronizer-l1/etherman"
 	"github.com/0xPolygonHermez/zkevm-synchronizer-l1/log"
+	storage "github.com/0xPolygonHermez/zkevm-synchronizer-l1/storage"
 	syncconfig "github.com/0xPolygonHermez/zkevm-synchronizer-l1/synchronizer/config"
 	"github.com/mitchellh/mapstructure"
 	"github.com/spf13/viper"
