--- db/interface.go
@@ -1,1 +1,0 @@
-package storage
