+++ state/mocks/storage_l1_info_tree_interface.go
@@ -9,7 +9,7 @@ import (
 
 	mock "github.com/stretchr/testify/mock"
 
-	pgstorage "github.com/0xPolygonHermez/zkevm-synchronizer-l1/db/pgstorage"
+	pgstorage "github.com/0xPolygonHermez/zkevm-synchronizer-l1/storage/pgstorage"
 
 	pgx "github.com/jackc/pgx/v4"
 )
