+++ storage/pgstorage/errors.go
@@ -1,0 +1,14 @@
+package pgstorage
+
+import "errors"
+
+var (
+	// ErrNotFound indicates the state database may be empty
+	ErrNotFound = errors.New("not found in db")
+	// ErrStorageNotFound is used when the object is not found in the storage
+	ErrStorageNotFound = errors.New("not found in the Storage")
+	// ErrStorageNotRegister is used when the object is not found in the synchronizer
+	ErrStorageNotRegister = errors.New("not registered storage")
+	// ErrNilDBTransaction indicates the db transaction has not been properly initialized
+	ErrNilDBTransaction = errors.New("database transaction not properly initialized")
+)
