+++ storage/pgstorage/pgstorage.go
@@ -1,0 +1,72 @@
+package pgstorage
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/0xPolygonHermez/zkevm-synchronizer-l1/log"
+
+	"github.com/jackc/pgx/v4"
+	"github.com/jackc/pgx/v4/pgxpool"
+)
+
+// PostgresStorage implements the Storage interface.
+type PostgresStorage struct {
+	*pgxpool.Pool
+}
+
+// getExecQuerier determines which execQuerier to use, dbTx or the main pgxpool
+func (p *PostgresStorage) getExecQuerier(dbTx pgx.Tx) execQuerier {
+	if dbTx != nil {
+		return dbTx
+	}
+	return p
+}
+
+// NewPostgresStorage creates a new Storage DB
+func NewPostgresStorage(cfg Config) (*PostgresStorage, error) {
+	log.Infof("Running DB migrations")
+	err := RunMigrationsUp(cfg)
+	if err != nil {
+		log.Errorf("Error executing migrations: %v", err)
+		return nil, err
+	}
+
+	config, err := pgxpool.ParseConfig(fmt.Sprintf("postgres://%s:%s@%s:%s/%s?pool_max_conns=%d", cfg.User, cfg.Password, cfg.Host, cfg.Port, cfg.Name, cfg.MaxConns))
+	if err != nil {
+		log.Errorf("Unable to parse DB config: %v\n", err)
+		return nil, err
+	}
+	db, err := pgxpool.ConnectConfig(context.Background(), config)
+	if err != nil {
+		log.Errorf("Unable to connect to database: %v\n", err)
+		return nil, err
+	}
+	return &PostgresStorage{db}, nil
+}
+
+// Rollback rollbacks a db transaction.
+func (p *PostgresStorage) Rollback(ctx context.Context, dbTx pgx.Tx) error {
+	if dbTx != nil {
+		return dbTx.Rollback(ctx)
+	}
+
+	return ErrNilDBTransaction
+}
+
+// Commit commits a db transaction.
+func (p *PostgresStorage) Commit(ctx context.Context, dbTx pgx.Tx) error {
+	if dbTx != nil {
+		return dbTx.Commit(ctx)
+	}
+	return ErrNilDBTransaction
+}
+
+// BeginDBTransaction starts a transaction block.
+func (p *PostgresStorage) BeginDBTransaction(ctx context.Context) (pgx.Tx, error) {
+	return p.Begin(ctx)
+}
+
+func (p *PostgresStorage) BeginStateTransaction(ctx context.Context) (pgx.Tx, error) {
+	return p.BeginDBTransaction(ctx)
+}
