+++ synchronizer/actions/etrog/processor_l1_sequence_batches.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"time"
 
-	"github.com/0xPolygonHermez/zkevm-synchronizer-l1/db/pgstorage"
 	"github.com/0xPolygonHermez/zkevm-synchronizer-l1/etherman"
 	"github.com/0xPolygonHermez/zkevm-synchronizer-l1/log"
+	"github.com/0xPolygonHermez/zkevm-synchronizer-l1/storage/pgstorage"
 	"github.com/0xPolygonHermez/zkevm-synchronizer-l1/synchronizer/actions"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/jackc/pgx/v4"
